requirements.txt
python-decouple==3.8
requests==2.28.1
twilio==7.16.4
